vehicleTracking/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
deep-sort-realtime==1.3.2
exceptiongroup==1.2.2
filelock==3.13.1
fire==0.7.0
fonttools==4.57.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.25.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
paddlepaddle-gpu==3.0.0
pandas==2.2.3
pillow==11.0.0
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
sympy==1.13.1
termcolor==3.0.1
tifffile==2025.3.30
tk==0.1.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.101
ultralytics-thop==2.0.14
urllib3==2.3.0
